requirements.txt
argcomplete==3.1.1
bcrypt==4.0.1
blinker==1.6.2
boto3==1.26.140
botocore==1.29.140
CacheControl==0.12.11
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
cfn-flip==1.3.0
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
dnspython==2.3.0
durationpy==0.5
ecdsa==0.18.0
firebase-admin==6.1.0
Flask==2.3.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-PyMongo==2.3.0
google-api-core==2.11.0
google-api-python-client==2.87.0
google-auth==2.18.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpcio==1.54.2
grpcio-status==1.54.2
gunicorn==20.1.0
hjson==3.1.0
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
kappa==0.6.0
MarkupSafe==2.1.2
msgpack==1.0.5
placebo==0.9.0
proto-plus==1.22.2
protobuf==4.23.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pymongo==4.3.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-slugify==8.0.1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
s3transfer==0.6.1
six==1.16.0
text-unidecode==1.3
toml==0.10.2
tqdm==4.66.1
troposphere==4.4.1
uritemplate==4.1.1
urllib3==1.26.16
Werkzeug==2.3.4
zappa==0.57.0
